Added working database
