+++ .env
@@ -1,0 +1,3 @@
+DB_HOST=localhost
+DB_PORT=5432
+DB_NAME=planet_of_the_shapes
