+++ create-db.sh
@@ -1,0 +1,3 @@
+createdb planet_of_the_shapes
+psql -f schema.sql planet_of_the_shapes
+psql -f seed.sql planet_of_the_shapes
