+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
+    "db:reset": "dropdb planet_of_the_shapes; bash create-db.sh",
     "start": "node ./bin/www"
   },
   "dependencies": {
@@ -14,10 +15,10 @@
     "express": "~4.16.1",
     "express-session": "^1.16.2",
     "http-errors": "~1.6.3",
+    "loading-spinner": "^1.2.1",
     "morgan": "~1.9.1",
     "pg-promise": "^9.0.3",
     "request": "^2.88.0",
     "socket.io": "^2.2.0"
-
   }
 }
