+++ public/Loading Screen/loading.ejs
@@ -1,0 +1,19 @@
+var loadingSpinner = require('loading-spinner');
+
+var dary = function() {
+  loadingSpinner.stop();
+
+  process.stdout.write('DA-RY !');
+};
+
+var legend = function() {
+  process.stdout.write('It\'s gonna be LE-GEN... Wait for it... ');
+
+  loadingSpinner.start(100, {
+    clearChar: true
+  });
+
+  setTimeout(dary, 1000);
+};
+
+legend();
