+++ public/stylesheets/login.css
@@ -1,0 +1,4 @@
+body{
+    background-image: url("../images/Backgroun2.png.jpg");
+    background-color: #cccccc;
+}
