+++ views/login.html
@@ -1,7 +1,15 @@
-<% include head %>
-<link rel="stylesheet" href="/stylesheets/login.css">
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Planet of The Shapes</title>
+    <link rel="stylesheet" href="login.css">
 </head>
-<form action="/login" method="post">
+
+<form action="/users/loginProcess" method="post">
     <section>
         <div class="container">
           <div class="row justify-content-center">
