+++ views/register.ejs
@@ -1,4 +1,4 @@
-<form action="/users/registerProcess" method="POST">
+<form action="/register" method="POST">
     <section>
         <div class="container">
           <div class="row justify-content-center">
@@ -12,11 +12,6 @@
               <div class="row align-items-center">
                 <div class="col mt-4">
                   <input type="text" name= "displayname" class="form-control" placeholder="Display Name">
-                </div>
-              </div>
-              <div class="row align-items-center mt-4">
-                <div class="col">
-                  <input type="email" name="email" class="form-control" placeholder="Email">
                 </div>
               </div>
               <div class="row align-items-center mt-4">
